+++ backend/src/neuroagent/app/app_utils.py
@@ -4,6 +4,8 @@ import logging
 from typing import Any
 
 from fastapi import HTTPException
+from pydantic import BaseModel, ConfigDict, Field
+from redis import asyncio as aioredis
 from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy.ext.asyncio import AsyncEngine, create_async_engine
 from starlette.status import HTTP_401_UNAUTHORIZED
@@ -11,6 +13,16 @@ from starlette.status import HTTP_401_UNAUTHORIZED
 from neuroagent.app.config import Settings
 
 logger = logging.getLogger(__name__)
+
+
+class RateLimitHeaders(BaseModel):
+    """Headers for the rate limits."""
+
+    x_ratelimit_limit: str = Field(alias="x-ratelimit-limit")
+    x_ratelimit_remaining: str = Field(alias="x-ratelimit-remaining")
+    x_ratelimit_reset: str = Field(alias="x-ratelimit-reset")
+
+    model_config = ConfigDict(validate_by_name=True, validate_by_alias=True)
 
 
 def setup_engine(
@@ -72,3 +84,73 @@ def validate_project(
     else:
         # No vlab nor project provided, nothing to do.
         return
+
+
+async def rate_limit(
+    redis_client: aioredis.Redis | None,
+    route_path: str,
+    limit: int,
+    expiry: int,
+    user_sub: str,
+) -> tuple[RateLimitHeaders, bool]:
+    """Check rate limiting for a given route and user.
+
+    Parameters
+    ----------
+    redis_client : aioredis.Redis
+        Redis client instance
+    route_path : str
+        Path of the route being accessed
+    limit : int
+        Maximum number of requests allowed
+    expiry : int
+        Time in seconds before the rate limit resets
+    user_sub : str
+        User identifier
+
+    Returns
+    -------
+    RateLimitHeaders
+        Pydantic class detailing rate limit info and meant to be dumped in response headers.
+    rate_limited
+        Whether the user is rate limited. In parent endpoint raise error if True.
+    """
+    if redis_client is None:
+        return RateLimitHeaders(
+            x_ratelimit_limit="-1", x_ratelimit_remaining="-1", x_ratelimit_reset="-1"
+        ), False  # redis disabled
+
+    # Create key using normalized route path and user sub
+    key = f"rate_limit:{user_sub}:{route_path}"
+
+    # Get current count
+    current = await redis_client.get(key)
+    current = int(current) if current else 0
+
+    if current > 0:
+        # Get the remaining time
+        ttl = await redis_client.pttl(key)
+        if current + 1 > limit:
+            # Rate limited
+            return RateLimitHeaders(
+                x_ratelimit_limit=str(limit),
+                x_ratelimit_remaining="0",
+                x_ratelimit_reset=str(round(ttl / 1000)),
+            ), True
+
+        # Not rate limited
+        await redis_client.incr(key)
+        return RateLimitHeaders(
+            x_ratelimit_limit=str(limit),
+            x_ratelimit_remaining=str(limit - current - 1),
+            x_ratelimit_reset=str(round(ttl / 1000)),
+        ), False
+
+    # Key did not exist yet
+    else:
+        await redis_client.set(key, 1, ex=expiry)
+        return RateLimitHeaders(
+            x_ratelimit_limit=str(limit),
+            x_ratelimit_remaining=str(limit - current - 1),
+            x_ratelimit_reset=str(expiry),
+        ), False