+++ backend/src/neuroagent/tools/web_search.py
@@ -1,6 +1,5 @@
 """Generic websearch tool."""
 
-import json
 from typing import ClassVar, Literal
 
 from pydantic import BaseModel, Field, SecretStr
@@ -27,6 +26,7 @@ class WebSearchInput(BaseModel):
     )
     max_results: int = Field(
         default=5,
+        ge=1,
         le=20,
         description="The maximum number of search results to return.",
     )
@@ -38,6 +38,24 @@ class WebSearchInput(BaseModel):
         default=None,
         description="A list of website to specifically exclude from the search. If the user specifies sources to avoid, specify it here.",
     )
+
+
+class SingleSearchOutput(BaseModel):
+    """Single search result."""
+
+    title: str
+    url: str
+    content: str
+    score: float
+    raw_content: str | None
+
+
+class WebSearchToolOutput(BaseModel):
+    """Output of the Web Search tool."""
+
+    query: str
+    results: list[SingleSearchOutput]
+    response_time: float
 
 
 class WebSearchTool(BaseTool):
@@ -52,7 +70,7 @@ class WebSearchTool(BaseTool):
     metadata: WebSearchMetadata
     input_schema: WebSearchInput
 
-    async def arun(self) -> str:
+    async def arun(self) -> WebSearchToolOutput:
         """Run the tools."""
         client = AsyncTavilyClient(
             api_key=self.metadata.tavily_api_key.get_secret_value()
@@ -64,7 +82,20 @@ class WebSearchTool(BaseTool):
             exclude_domains=self.input_schema.exclude_domains,
             time_range=self.input_schema.time_range,
         )
-        return json.dumps(search_result)
+        return WebSearchToolOutput(
+            query=search_result["query"],
+            response_time=search_result["response_time"],
+            results=[
+                SingleSearchOutput(
+                    title=result["title"],
+                    url=result["url"],
+                    content=result["content"],
+                    score=result["score"],
+                    raw_content=result.get("raw_content"),
+                )
+                for result in search_result["results"]
+            ],
+        )
 
     @classmethod
     async def is_online(cls) -> bool: