+++ backend/tests/tools/test_literature_search_tool.py
@@ -1,28 +1,30 @@
 """Tests Literature Search tool."""
 
-import json
-
 import httpx
 import pytest
 
 from neuroagent.tools import LiteratureSearchTool
 from neuroagent.tools.literature_search_tool import (
+    ArticleOutput,
     LiteratureSearchInput,
     LiteratureSearchMetadata,
+    LiteratureSearchToolOutput,
+    ParagraphOutput,
 )
 
 
 class TestLiteratureSearchTool:
     @pytest.mark.asyncio
     async def test_arun(self, httpx_mock):
-        url = "http://fake_url?query=covid+19&retriever_k=100&use_reranker=true&reranker_k=5"
+        url = "http://fake_url?query=covid+19&retriever_k=100&use_reranker=true&reranker_k=100"
         reranker_k = 5
+        retriever_k = 100
 
         fake_response = [
             {
                 "article_title": "great_title",
                 "article_authors": ["author1", "author2"],
-                "article_id": "super_id",
+                "article_id": f"super_id_{i % 30}",
                 "article_doi": "magnigficent_doi",
                 "pubmed_id": "random_pubmed_id",
                 "date": "02/02/2022",
@@ -37,7 +39,7 @@ class TestLiteratureSearchTool:
                 "section": "Introduction",
                 "context_id": "21",
             }
-            for _ in range(reranker_k)
+            for i in range(retriever_k)
         ]
 
         httpx_mock.add_response(
@@ -46,33 +48,33 @@ class TestLiteratureSearchTool:
         )
 
         tool = LiteratureSearchTool(
-            input_schema=LiteratureSearchInput(query="covid 19"),
+            input_schema=LiteratureSearchInput(
+                query="covid 19", article_number=reranker_k
+            ),
             metadata=LiteratureSearchMetadata(
                 literature_search_url=url,
                 httpx_client=httpx.AsyncClient(),
                 token="fake_token",
-                retriever_k=100,
+                retriever_k=retriever_k,
                 use_reranker=True,
-                reranker_k=reranker_k,
             ),
         )
         response = await tool.arun()
-        assert isinstance(response, str)
-        response = json.loads(response)
-        assert len(response) == reranker_k
-        assert isinstance(response[0], dict)
+        assert isinstance(response, LiteratureSearchToolOutput)
+        assert len(response.articles) == reranker_k
+        assert isinstance(response.articles[0], ArticleOutput)
+        assert isinstance(response.articles[0].paragraphs[0], ParagraphOutput)
 
 
 class TestCreateQuery:
     tool = LiteratureSearchTool(
-        input_schema=LiteratureSearchInput(query="covid 19"),
+        input_schema=LiteratureSearchInput(query="covid 19", article_number=1),
         metadata=LiteratureSearchMetadata(
             literature_search_url="https://fake_url.com",
             httpx_client=httpx.AsyncClient(),
             token="fake_token",
             retriever_k=100,
             use_reranker=False,
-            reranker_k=1,
         ),
     )