+++ backend/tests/tools/test_traces_tool.py
@@ -8,7 +8,12 @@ import httpx
 import pytest
 
 from neuroagent.tools import GetTracesTool
-from neuroagent.tools.traces_tool import GetTracesInput, GetTracesMetadata
+from neuroagent.tools.traces_tool import (
+    GetTracesInput,
+    GetTracesMetadata,
+    GetTracesToolOutput,
+    Trace,
+)
 
 
 class TestTracesTool:
@@ -56,11 +61,9 @@ class TestTracesTool:
             Bucket="test-bucket", Key="test-br-key"
         )
 
-        assert isinstance(response, str)
-        response = json.loads(response)
-        assert len(response) == 2
-        assert isinstance(response[0], dict)
-        assert isinstance(response[0], dict)
+        assert isinstance(response, GetTracesToolOutput)
+        assert len(response.traces) == 2
+        assert isinstance(response.traces[0], Trace)
 
     @pytest.mark.httpx_mock(can_send_already_matched_responses=True)
     @pytest.mark.asyncio
@@ -107,10 +110,9 @@ class TestTracesTool:
             Bucket="test-bucket", Key="test-br-key"
         )
 
-        assert isinstance(response, str)
-        response = json.loads(response)
-        assert len(response) == 2
-        assert isinstance(response[0], dict)
+        assert isinstance(response, GetTracesToolOutput)
+        assert len(response.traces) == 2
+        assert isinstance(response.traces[0], Trace)
 
     @pytest.mark.asyncio
     async def test_arun_errors(self, httpx_mock, brain_region_json_path):